charts drawer added floats changed to doubles

+++ zad1/src/CRandom/Random.java
@@ -3,11 +3,11 @@ package CRandom;
 public class Random {
     private static java.util.Random r = new java.util.Random();
 
-    public static float random(float min, float max) {
+    public static double random(double min, double max) {
         return r.nextFloat() * (max - min) + min;
     }
 
-    public static float randomGaussian(float min, float max) {
-        return (float) r.nextGaussian() * (max - min) + min;
+    public static double randomGaussian(double min, double max) {
+        return r.nextGaussian() * (max - min) + min;
     }
 }
