+++ zad1/src/Signals/NoiseGaussian.java
@@ -17,8 +17,8 @@ public class NoiseGaussian extends Signal {
     }
 
     @Override
-    public Map<Float, Float> generate(float fs) {
-        Map<Float, Float> chart = new HashMap<>();
+    public Map<Double, Double> generate(float fs) {
+        Map<Double, Double> chart = new HashMap<>();
 
         float tx= t1+d;
         float Ts = 1/fs;
