+++ zad1/src/Signals/SignalRectangularSymmetric.java
@@ -19,8 +19,8 @@ public class SignalRectangularSymmetric extends Signal {
     }
 
     @Override
-    public Map<Float, Float> generate(float fs) {
-        Map<Float, Float> chart = new HashMap<>();
+    public Map<Double, Double> generate(float fs) {
+        Map<Double, Double> chart = new HashMap<>();
 
         return chart;
     }
